refactor : 글로벌 유틸 로직 패키지 구조화
